+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/default-o-table.datasource.ts
@@ -68,8 +68,8 @@ export class DefaultOTableDataSource extends DataSource<any> implements OTableDa
       this._paginator = table.matpaginator;
     }
 
-    if (this.table.activeVirtualScroll) {
-      this.table.scrollStrategy.renderedRangeStream
+    if (this.table.virtualScrollViewport) {
+      this.table.virtualScrollViewport.renderedRangeStream
         .pipe(distinctUntilChanged())
         .subscribe(
           (value: ListRange) => {
@@ -119,7 +119,7 @@ export class DefaultOTableDataSource extends DataSource<any> implements OTableDa
       }
     }
 
-    if (this.table.activeVirtualScroll) {
+    if (this.table.virtualScrollViewport) {
       displayDataChanges.push(this._virtualPageChange);
     }
 
@@ -172,7 +172,7 @@ export class DefaultOTableDataSource extends DataSource<any> implements OTableDa
             when the data is very large, the application crashes so it gets a limited range of data the first time
             because at next the CustomVirtualScrollStrategy will emit event OnRangeChangeVirtualScroll
           */
-          if (this.table.activeVirtualScroll && !this._paginator) {
+          if (this.table.virtualScrollViewport && !this._paginator) {
             data = this.getVirtualScrollData(data, new OnRangeChangeVirtualScroll({ start: 0, end: Codes.LIMIT_SCROLLVIRTUAL }));
           }