+++ projects/ontimize-web-ngx/src/lib/interfaces/o-form-layout-manager-mode.interface.ts
@@ -1,3 +1,4 @@
+import { Observable } from 'rxjs';
 import { FormLayoutDetailComponentData } from '../types/form-layout-detail-component-data.type';
 import { ILayoutManagerComponent } from './layout-manager-component.interface';
 
@@ -7,12 +8,12 @@ export interface OFormLayoutManagerMode {
   initializeComponentState?: (state: any) => void;
   getDataToStore: () => object;
   getFormCacheData: () => FormLayoutDetailComponentData;
-  setModifiedState: (modified: boolean) => void;
+  setModifiedState: (formAttr: string, modified: boolean, confirmExit: boolean) => void;
   updateNavigation: (data: any, keysValues: any, insertionMode?: boolean) => void;
   updateActiveData: (data: any) => void;
   getRouteOfActiveItem: () => any[];
   isMainComponent: (comp: ILayoutManagerComponent) => boolean;
   openDetail?: (detail: FormLayoutDetailComponentData) => void;
   closeDetail: () => void;
-  canAddDetailComponent: () => boolean;
+  canAddDetailComponent: () => boolean | Observable<boolean>;
 }